IOS-7350 Move to StakeKitMapper
