+++ Tangem/Modules/Send/StakingModel.swift
@@ -109,7 +109,7 @@ private extension StakingModel {
             .withWeakCaptureOf(self)
             .flatMap { model, transaction in
                 model.sendTransactionDispatcher
-                    .sendPublisher(transaction: .staking(model.mapToStakeKitTransaction(transaction)))
+                    .sendPublisher(transaction: .staking(StakeKitMapper().mapToStakeKitTransaction(transaction)))
             }
             .handleEvents(receiveOutput: { [weak self] output in
                 self?.proceed(result: output)
@@ -132,18 +132,6 @@ private extension StakingModel {
         case .success:
             _transactionTime.send(Date())
         }
-    }
-
-    // TODO: get fee, amount and source address
-    private func mapToStakeKitTransaction(_ transaction: StakingTransactionInfo) -> StakeKitTransaction {
-        let stakeKitTransaction = StakeKitTransaction(
-            amount: Amount(type: .coin, currencySymbol: "", value: 0, decimals: 0),
-            fee: Fee(Amount(type: .coin, currencySymbol: "", value: 0, decimals: 0)),
-            sourceAddress: "",
-            unsignedData: transaction.unsignedTransactionData
-        )
-
-        return stakeKitTransaction
     }
 }
